# services/graph-service/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "liner-notes API",
    "description": "Graph-driven vinyl record collection explorer",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/explore/musician/{name}": {
      "get": {
        "summary": "Releases featuring this musician, with instrument and role on each result",
        "tags": ["explore"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["discogsId", "title"],
                    "properties": {
                      "discogsId": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "artist": {
                        "type": "string",
                        "nullable": true
                      },
                      "pressingYear": {
                        "type": "integer",
                        "nullable": true
                      },
                      "format": {
                        "type": "string",
                        "nullable": true
                      },
                      "thumbUrl": {
                        "type": "string",
                        "nullable": true
                      },
                      "instrument": {
                        "type": "string",
                        "nullable": true
                      },
                      "role": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explore/producer/{name}": {
      "get": {
        "summary": "Releases this person is credited on as a producer",
        "tags": ["explore"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["discogsId", "title"],
                    "properties": {
                      "discogsId": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "artist": {
                        "type": "string",
                        "nullable": true
                      },
                      "pressingYear": {
                        "type": "integer",
                        "nullable": true
                      },
                      "format": {
                        "type": "string",
                        "nullable": true
                      },
                      "thumbUrl": {
                        "type": "string",
                        "nullable": true
                      },
                      "instrument": {
                        "type": "string",
                        "nullable": true
                      },
                      "role": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explore/engineer/{name}": {
      "get": {
        "summary": "Releases this person is credited on as an engineer",
        "tags": ["explore"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["discogsId", "title"],
                    "properties": {
                      "discogsId": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "artist": {
                        "type": "string",
                        "nullable": true
                      },
                      "pressingYear": {
                        "type": "integer",
                        "nullable": true
                      },
                      "format": {
                        "type": "string",
                        "nullable": true
                      },
                      "thumbUrl": {
                        "type": "string",
                        "nullable": true
                      },
                      "instrument": {
                        "type": "string",
                        "nullable": true
                      },
                      "role": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explore/studio/{name}": {
      "get": {
        "summary": "Releases recorded at this studio",
        "tags": ["explore"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["discogsId", "title"],
                    "properties": {
                      "discogsId": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "artist": {
                        "type": "string",
                        "nullable": true
                      },
                      "pressingYear": {
                        "type": "integer",
                        "nullable": true
                      },
                      "format": {
                        "type": "string",
                        "nullable": true
                      },
                      "thumbUrl": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explore/label/{name}": {
      "get": {
        "summary": "Releases on this label",
        "tags": ["explore"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["discogsId", "title"],
                    "properties": {
                      "discogsId": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "artist": {
                        "type": "string",
                        "nullable": true
                      },
                      "pressingYear": {
                        "type": "integer",
                        "nullable": true
                      },
                      "format": {
                        "type": "string",
                        "nullable": true
                      },
                      "thumbUrl": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explore/genre/{name}": {
      "get": {
        "summary": "Releases in this genre",
        "tags": ["explore"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["discogsId", "title"],
                    "properties": {
                      "discogsId": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "artist": {
                        "type": "string",
                        "nullable": true
                      },
                      "pressingYear": {
                        "type": "integer",
                        "nullable": true
                      },
                      "format": {
                        "type": "string",
                        "nullable": true
                      },
                      "thumbUrl": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explore/style/{name}": {
      "get": {
        "summary": "Releases in this style",
        "tags": ["explore"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["discogsId", "title"],
                    "properties": {
                      "discogsId": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "artist": {
                        "type": "string",
                        "nullable": true
                      },
                      "pressingYear": {
                        "type": "integer",
                        "nullable": true
                      },
                      "format": {
                        "type": "string",
                        "nullable": true
                      },
                      "thumbUrl": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}